requirements.txt
boto3>=1.20.13,<1.21
python-dotenv>=0.19.2,<0.20
beautifulsoup4>=4.10.0,<4.11
jinja2>=3.0.3,<3.1